requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2021.10.8        
charset-normalizer==2.0.12
dj-database-url==0.5.0    
Django==4.0.3
django-anymail==8.5
django-autoslug==1.9.8
django-crispy-forms==1.14.0
django-environ==0.8.1
django-model-utils==4.2.0
django-on-heroku==1.1.2
django-resized==0.3.11
gunicorn==20.0.4
heroku==0.1.4
idna==3.3
Pillow==9.1.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
python-dateutil==1.5
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
stripe==2.70.0
toml==0.10.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0

